requirements.txt
# Remove pysqlite3-binary as it's not available
# pysqlite3-binary
numpy==1.26.4
librosa==0.10.2.post1
scikit-learn==1.5.2
matplotlib==3.8.0
tf-keras
tensorflow==2.17.1
soundfile==0.12.1
streamlit==1.40.1
openpyxl==3.1.2
chromadb==0.5.18
sentence-transformers==3.3.0
openai==1.3.0
httpx>=0.24.1